requirements.txt
allure-pytest
pre-commit
pytest
requests
waiting
